+++ app_bot/bot/core/log.py
@@ -15,4 +15,7 @@ def setup_logging(suppress_asyncio: bool = True, suppress_urllib3: bool = True) 
         logging.getLogger('urllib3').setLevel(logging.WARNING)
 
     logging.getLogger('pyrogram').setLevel(logging.WARNING)
-    logging.getLogger('sqlalchemy.engine.Engine').handlers.pop()
+    try:
+        logging.getLogger('sqlalchemy.engine.Engine').handlers.pop()
+    except IndexError:
+        pass